Add option to build LLVM from source in CMake
Allow LLVM to be built from source as part of the same build. This is
convenient because we can quickly modify LLVM and run with the changes,
and LLVM will inherit the build type used for triton (i.e. release vs
debug) so we can debug through LLVM code.

With this change, the triton debug build can be configured as follows:
```
cmake -S triton -B build_dbg -GNinja -DCMAKE_BUILD_TYPE=Debug \
    -DTRITON_CODEGEN_BACKENDS="nvidia;amd" \
    -DTRITON_LLVM_SOURCE_DIR="/Users/neildhar/triton_llvm"
```