Use libstdc++ from system toolchain for AlmaLinux
The AlmaLinux build currently builds against libstdc++ from GCC 14,
because installing clang also pulls in GCC 14, and clang automatically
picks the newest libstdc++ on the system.

Instead, pointing clang to the system GCC installation should link
against the corresponding libstdc++. This improves the portability of
the generated binaries. The AlmaLinux binaries should now be compatible
with a superset of the systems that the Ubuntu binaries are compatible
with.

We can then separately evaluate removing the Ubuntu build if this works
for all users.
(I assume these have to be separate changes anyway for CI)